api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tenniscoach-api"
version = "0.1.0"
description = "テニス記録アプリ AI コーチング API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "google-genai>=1.0.0",
    "pydantic>=2.0.0",
    "toml>=0.10.0",
    "tenacity>=9.0.0",
    "google-api-core>=2.30.0",
    "python-dotenv>=1.2.2",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pytest>=8.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/tenniscoach_api"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "SIM", "RUF"]

[tool.mypy]
python_version = "3.12"
strict = true

[[tool.mypy.overrides]]
module = "toml"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "google.api_core.*"
ignore_missing_imports = true
